test(integration): move builtin dang regression into its own file
